# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Point of Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Brands
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List all brands
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      summary: Create a brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    patch:
      operationId: updateBrand
      tags: [brands]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
    delete:
      operationId: deleteBrand
      tags: [brands]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Purchases
...
  /dashboard/recent-sales:
    get:
      operationId: getRecentSales
      tags: [dashboard]
      summary: Recent sales for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recent sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleWithDetails"
  /dashboard/low-stock:
    get:
      operationId: getLowStockProducts
      tags: [dashboard]
      summary: Products with low or zero stock
      parameters:
        - name: threshold
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Low stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWithStock"
  /dashboard/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [dashboard]
      summary: Daily sales for the past N days
      parameters:
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Sales trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesTrendPoint"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Brand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateBrandBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Product:
      type: object
      properties:
        id:
          type: integer
        productCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        barcode:
          type: ["string", "null"]
        salesPrice:
          type: number
        cost:
          type: number
        discount:
          type: number
        discountType:
          type: string
          enum: [fixed, percentage]
        hasExpiry:
          type: boolean
        valuationMethod:
          type: string
          enum: [fifo, lifo, average]
        isActive:
          type: boolean
        categoryName:
          type: ["string", "null"]
        brandName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productCode
        - name
        - salesPrice
        - cost
        - discount
        - discountType
        - hasExpiry
        - valuationMethod
        - isActive
        - createdAt
    ProductWithStock:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            currentStock:
              type: number
          required:
            - currentStock
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactNo:
          type: ["string", "null"]
        creditLimit:
          type: number
        loyaltyPoints:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - creditLimit
        - loyaltyPoints
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactNo:
          type: ["string", "null"]
        creditLimit:
          type: number
        loyaltyPoints:
          type: number
      required:
        - name
        - creditLimit
...
    SaleDetail:
      type: object
      properties:
        id:
          type: integer
        saleId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        salesPrice:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - id
        - saleId
        - productId
        - productName
        - productCode
        - salesPrice
        - discount
        - qty
    PaymentTransaction:
      type: object
      properties:
        id:
          type: integer
        saleId:
          type: integer
        paymentType:
          type: string
          enum: [cash, cheque, card, voucher, loyalty]
        amount:
          type: number
      required:
        - id
        - saleId
        - paymentType
        - amount
    Sale:
      type: object
      properties:
        id:
          type: integer
        saleDate:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        billNo:
          type: string
        totalAmount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - saleDate
        - billNo
        - totalAmount
        - createdAt
    SaleWithDetails:
      allOf:
        - $ref: "#/components/schemas/Sale"
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: "#/components/schemas/SaleDetail"
            payments:
              type: array
              items:
                $ref: "#/components/schemas/PaymentTransaction"
          required:
            - details
            - payments
...
    SalesTrendPoint:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        orderCount:
          type: integer
      required:
        - date
        - totalSales
        - orderCount
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        totalQtySold:
          type: number
        totalRevenue:
          type: number
      required:
        - productId
        - productName
        - productCode
        - totalQtySold
        - totalRevenue