# lib/api-zod/src/generated/types/paymentTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Point of Sales System API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentTransactionPaymentType } from "./paymentTransactionPaymentType";

export interface PaymentTransaction {
  id: number;
  saleId: number;
  paymentType: PaymentTransactionPaymentType;
  amount: number;
}